requirements.txt
async-timeout==5.0.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
cs50==9.4.0
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.10.1
Flask-Session==0.8.0
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
ordered-set==4.1.0
packaging==24.2
Pygments==2.19.1
python-dotenv==1.0.1
redis==5.2.1
rich==13.9.4
SQLAlchemy==2.0.37
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.2
